feat: add working user controller Implemented custom exceptions, DTOs and MapStruct with mapping tests. Changed User and System entity names to not collide with built-in classes.

+++ src/main/java/pl/polsl/softhouse/services/UserService.java
@@ -1,13 +1,16 @@
 package pl.polsl.softhouse.services;
 
 import java.util.List;
-import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 
+import pl.polsl.softhouse.dto.user.UserAuthDto;
 import pl.polsl.softhouse.dto.user.UserDto;
+import pl.polsl.softhouse.dto.user.UserInfoDto;
 import pl.polsl.softhouse.dto.user.UserMapper;
-import pl.polsl.softhouse.entities.User;
+import pl.polsl.softhouse.entities.UserEntity;
+import pl.polsl.softhouse.exceptions.InvalidDataException;
 import pl.polsl.softhouse.exceptions.user.UserAlreadyExistsException;
 import pl.polsl.softhouse.exceptions.user.UserNotFoundException;
 import pl.polsl.softhouse.repositories.UserRepository;
@@ -24,57 +27,72 @@ public class UserService {
         this.userMapper = userMapper;
     }
 
-    public List<User> getUsers() {
-        return userRepository.findAll();
+    public List<UserInfoDto> getAllUsers() {
+        return userRepository.findAll()
+                .stream()
+                .map(userMapper::userToInfoDto)
+                .collect(Collectors.toList());
     }
 
-    public User getUserById(Long id) {
-        return userRepository.findById(id).get();
-    }
-
-    public User addUser(UserDto userDto) {
-        if (!checkIfUsernameUnique(userDto.getUsername())) {
-            throw new UserAlreadyExistsException("User already exists."); // TODO
+    public UserInfoDto getUserById(Long id) {
+        if (id == null) {
+            throw new InvalidDataException("No data sent."); // TODO.
         }
 
-        User user = userMapper.dtoToUser(userDto);
-        user.setId(0L);
-        user.setActive(true);
+        return userRepository.findById(id)
+                .map(userMapper::userToInfoDto)
+                .orElseThrow(() -> new UserNotFoundException(id)); // TODO
+    }
 
-        return userRepository.save(user);
+    public void addUser(UserDto userDto) {
+        if (userDto == null) {
+            throw new InvalidDataException("No data sent."); // TODO: Custom exception maybe.
+        }
+
+        if (!checkIfUsernameUnique(userDto.getUsername())) {
+            throw UserAlreadyExistsException.fromUsername(userDto.getUsername()); // TODO
+        }
+
+        UserEntity user = userMapper.createUserFromDto(userDto);
+
+        userRepository.save(user);
     }
 
     public void deleteUser(Long id) {
+        if (id == null) {
+            throw new InvalidDataException("No data sent."); // TODO: Custom exception maybe.
+        }
+
         if (!userRepository.existsById(id)) {
-            throw new UserNotFoundException("User does not exist."); // TODO
+            throw new UserNotFoundException(id); // TODO
         }
 
         userRepository.deleteById(id);
     }
 
-    // TODO: A lot of parameters, probably not good.
-    public User updateUser(Long id, UserDto userDto) {
-        Optional<User> foundUser = userRepository.findById(id);
-        if (foundUser.isEmpty()) {
-            throw new UserNotFoundException("User does not exist.");
+    // TODO: Test.
+    public void updateUser(Long id, UserDto userDto) {
+        if (id == null || userDto == null) {
+            throw new InvalidDataException("No data sent."); // TODO: Custom exception maybe.
         }
 
-        User user = userMapper.updateUser(userDto, foundUser.get());
-
-        return userRepository.save(user);
+        userRepository.findById(id)
+                .map(foundUser -> userMapper.updateUser(userDto, foundUser))
+                .map(userRepository::save)
+                .orElseThrow(() -> new UserNotFoundException(id));
     }
 
-    public User getUserByUsername(String username) {
-        Optional<User> foundUser = userRepository.findByUsername(username);
-        if (foundUser.isEmpty()) {
-            throw new UserNotFoundException("User does not exist.");
+    public UserAuthDto getUserAuthByUsername(String username) {
+        if (username == null) {
+            throw new InvalidDataException("No data sent."); // TODO: Custom exception maybe.
         }
 
-        return foundUser.get();
+        return userRepository.findByUsername(username)
+                .map(userMapper::userToAuthDto)
+                .orElseThrow(() -> UserNotFoundException.fromUsername(username));
     }
 
     private boolean checkIfUsernameUnique(String username) {
-        Optional<User> foundUser = userRepository.findByUsername(username);
-        return foundUser.isEmpty();
+        return userRepository.findByUsername(username).isEmpty();
     }
 }
